# docs/archive/openapi.monolith.yaml
openapi: 3.0.3
info:
  title: Appetitebind
  version: 0.1.0
  description: >-
    Enterprise AI risk operating system that binds every AI use case to risk
    appetite and a fairness policy, runs continuous control testing with a
    governed kill-switch, and predicts residual-risk threshold breaches before
    the board pack is written. Not a cyber SOC tool and not a model-validation
    factory — the RMF lifecycle for AI itself.
  x-product-slug: ai-cyber-appetitebind
  x-source: ai-in-cyber/deloitte-gx-ai-and-risk-management/
servers:
  - url: https://api.example.com
...
paths:
...
  /v1/control-tests:
    get:
      operationId: listAiControlTests
      tags: [Controls]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: useCaseId
          in: query
          schema:
            type: string
        - name: testType
          in: query
          schema:
            type: string
            enum: [out_of_sample, stress, customer_journey, data_bias, accuracy]
      responses:
        "200":
          description: Control tests page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeAiControlTest"
    post:
      operationId: scheduleAiControlTest
      tags: [Controls]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiControlTestCreate"
      responses:
        "201":
          description: Control test scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeAiControlTest"
  /v1/control-tests/{testId}/evidence:
    parameters:
      - $ref: "#/components/parameters/TestId"
    post:
      operationId: submitControlTestEvidence
      tags: [Controls]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestEvidence"
      responses:
        "201":
          description: Evidence captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeAiControlTest"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    TestId:
      name: testId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    AiControlTest:
      type: object
      required: [id, useCaseId, testType, status]
      properties:
        id:
          type: string
        useCaseId:
          type: string
        testType:
          type: string
          enum: [out_of_sample, stress, customer_journey, data_bias, accuracy]
        status:
          type: string
          enum: [scheduled, running, passed, failed]
        frequency:
          type: string
        lastEvidenceAt:
          type: string
          format: date-time
    AiControlTestCreate:
      type: object
      required: [useCaseId, testType]
      properties:
        useCaseId:
          type: string
        testType:
          type: string
          enum: [out_of_sample, stress, customer_journey, data_bias, accuracy]
        frequency:
          type: string
        ownerId:
          type: string
    TestEvidence:
      type: object
      required: [outcome]
      properties:
        outcome:
          type: string
          enum: [passed, failed]
        evidenceRef:
          type: string
        detail:
          type: string
        sampleSize:
          type: integer
...
    DataEnvelopeAiControlTest:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AiControlTest"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeAiControlTest:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AiControlTest"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...