# platform/api-server/src/domains/controls/handlers/control-test.handlers.ts
/**
 * Handlers for ControlTest
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: controls
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ControlsDomainModule } from "../dependencies/controls-ddd.dependencies.js";
import type { ListAiControlTestsInput, ScheduleAiControlTestInput } from "@appetitebind/services/controls";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/control-tests
 * Operation: listAiControlTests
 */
export async function listAiControlTests(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], useCaseId?: string | string[], testType?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: ControlsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input = {
    ...(request.query ?? {}),
  } as ListAiControlTestsInput;
  const result = await deps.useCases.controlTests.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/control-tests
 * Operation: scheduleAiControlTest
 */
export async function scheduleAiControlTest(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: ControlsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input = {
    ...(request.body ?? {}),
  } as ScheduleAiControlTestInput;
  const result = await deps.useCases.controlTests.get.execute(input);
  return reply.code(201).send(result);
}