# platform/api-server/src/domains/controls/handlers/evidence.handlers.ts
/**
 * Handlers for Evidence
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: controls
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ControlsDomainModule } from "../dependencies/controls-ddd.dependencies.js";
import type { SubmitControlTestEvidenceInput } from "@appetitebind/services/controls";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for POST /v1/control-tests/{testId}/evidence
 * Operation: submitControlTestEvidence
 */
export async function submitControlTestEvidence(
  request: FastifyRequest<{ Params: { testId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: ControlsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input = {
    testId: request.params.testId,
    ...(request.body ?? {}),
  } as SubmitControlTestEvidenceInput;
  const result = await deps.useCases.evidences.create.execute(input);
  return reply.code(201).send(result);
}